backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
dlib-bin
face-recognition==1.3.0
Pillow==11.1.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
aiofiles==24.1.0
